feat: refactor products api, tests
